moved from alacritty to wezterm; added some language servers

+++ install.sh
@@ -16,27 +16,29 @@ _has() {
 # create directories if not exist
 mkdir -pv $HOME/.cache/zsh
 mkdir -pv $HOME/.config/alacritty
+mkdir -pv $HOME/.config/wezterm
 mkdir -pv $HOME/.ssh
 mkdir -pv $HOME/.zfunc
 
 # remove old stuff
 rm -Rf $HOME/.zplug
 
-
 # install software check linux or mac
 if [[ "$OSTYPE" == "linux-gnu"* ]]; then
     echo "Arch Linux"
     
-    # install yay
-    pacman -S --needed git base-devel && git clone https://aur.archlinux.org/paru.git && cd paru && makepkg -si && cd .. && rm -rf paru
+    # install paru (needs to build using rust)
+    if ! _has paru; then
+        pacman -S --needed git base-devel && git clone https://aur.archlinux.org/paru.git && cd paru && makepkg -si && cd .. && rm -rf paru
+    fi
     
     # install packages
-    sudo pacman -S tmux neovim curl gawk fzf nodejs exa starship
+    sudo pacman -S tmux neovim curl gawk fzf nodejs exa starship ripgrep fd neofetch
     
     # install GUI packages
     if [[ "$USE_GUI" == 1 ]]; then
         echo "GUI used..."
-        sudo pacman -S alacritty ttf-firacode-nerd-font adobe-source-code-pro-fonts
+        sudo pacman -S wezterm ttf-firacode-nerd-font adobe-source-code-pro-fonts
     fi
 
     # install miniconda
@@ -47,9 +49,11 @@ elif [[ "$OSTYPE" == "darwin"* ]]; then
     echo "Mac OSX"
     # install homebrew
     /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
-    brew install tmux neovim curl gawk fzf nodejs exa starship
+    brew install tmux neovim curl gawk fzf nodejs exa starship fd
     brew tap homebrew/cask-fonts
-    brew install --cask font-source-code-pro alacritty font-fira-mono-nerd-font
+    brew install --cask font-source-code-pro wezterm font-fira-mono-nerd-font
+    brew tap nidnogg/zeitfetch
+    brew install zeitfetch
 
     # Install miniconda
     curl https://repo.anaconda.com/miniconda/Miniconda3-latest-MacOSX-x86_64.sh -o ~/miniconda.sh
@@ -78,14 +82,21 @@ ln -s $HOME/dotfiles/zsh/zshrc.ecare $HOME/.zshrc
 rm -rf $HOME/.config/nvim
 ln -s $HOME/dotfiles/nvim $HOME/.config/nvim
 
-rm -f $HOME/.alacritty.yml
-ln -s $HOME/dotfiles/alacritty/alacritty.yml.macbook $HOME/.alacritty.yml
+rm -f $HOME/.alacritty.yml $HOME/.config/alacritty/alacritty.yml
+ln -s $HOME/dotfiles/alacritty/alacritty.yml.macbook $HOME/.config/alacritty/alacritty.yml
+rm -f $HOME/.wezterm.lua $HOME/.config/wezterm/wezterm.lua
+ln -s $HOME/dotfiles/wezterm/wezterm.lua $HOME/.config/wezterm/wezterm.lua
 
 rm -f $HOME/.tmux.conf
 ln -s $HOME/dotfiles/tmux/tmux.conf $HOME/.tmux.conf
 
 rm -f $HOME/.ssh/config
 ln -s ~/dotfiles/ssh/config ~/.ssh/config
+
+rm -f $HOME/.ignore
+rm -f $HOME/.gitconfig
+ln -s ~/dotfiles/git/ignore ~/.ignore
+ln -s ~/dotfiles/git/gitconfig ~/.gitconfig
 
 eval "$($HOME/miniconda/bin/conda shell.zsh hook)"
 conda config --set changeps1 False
